hf_deployment/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
transformers==4.45.0
torch>=2.6.0
torchvision>=0.19.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
gunicorn>=21.0.0